Split folder tree hit target widget
